Make helper functions compatible with torch.vmap.
Currently, `torch.vmap` does *not* support control-flow operators like `if`
statements, so we had to replace the if statement in `_CustomizedTorchLog`
with equivalent logic implemented via a`torch.where` map. Informal testing,
including training deep RNNs and estimating the spectrum of Lyapunov exponents
in parallel for all systems in https://github.com/GilpinLab/dysts, shows
that precision and performance of the updated helper functions is comparable
to that of the previous (non-vmappable) helper functions.